+++ src/rogallo/widgets/command_line/widget.py
@@ -29,6 +29,10 @@ from textual_enhanced.binding import HelpfulBinding
 from textual_enhanced.commands import Quit
 
 ##############################################################################
+# Wasat imports.
+from wasat import GeminiURI
+
+##############################################################################
 # Local imports.
 from ...data import Bookmarks, CommandLineHistory, LocationHistory, NavigationHistory
 from ...types import short_location
@@ -142,6 +146,8 @@ class CommandLine(Vertical):
     """The history of navigation through locations."""
     bookmarks: var[Bookmarks] = var(list)
     """The bookmarks for the application."""
+    known_hosts: var[list[GeminiURI]] = var(list)
+    """The hosts known to the application."""
 
     _input = query_one(Input)
     """The input widget for the command line."""
@@ -172,6 +178,7 @@ class CommandLine(Vertical):
                             (str(visit.location) for visit in self.location_history),
                             (str(visit) for visit in self.navigation_history),
                             (str(bookmark.location) for bookmark in self.bookmarks),
+                            (str(known_host) for known_host in self.known_hosts),
                             (
                                 short_location(visit.location)
                                 for visit in self.location_history
@@ -183,6 +190,10 @@ class CommandLine(Vertical):
                             (
                                 short_location(bookmark.location)
                                 for bookmark in self.bookmarks
+                            ),
+                            (
+                                short_location(known_host)
+                                for known_host in self.known_hosts
                             ),
                         )
                     )
@@ -216,6 +227,11 @@ class CommandLine(Vertical):
         command: str
         """The command that was executed."""
 
+    def _refresh_suggestions(self) -> None:
+        """Refresh the suggestions for the input."""
+        if self.is_mounted:
+            self._input.suggester = self._history_suggester
+
     def handle_input(self, command: str) -> None:
         """Handle input from the user.
 
@@ -229,7 +245,7 @@ class CommandLine(Vertical):
                 self.history.add(command)
                 self.post_message(self.CommandExecuted(self, command))
                 self._input.value = ""
-                self._input.suggester = self._history_suggester
+                self._refresh_suggestions()
                 return
         self.notify("Unable to handle that input", title="Error", severity="error")
 
@@ -245,8 +261,19 @@ class CommandLine(Vertical):
 
     def _watch_history(self) -> None:
         """React to history being updated."""
-        if self.is_mounted:
-            self._input.suggester = self._history_suggester
+        self._refresh_suggestions()
+
+    def _watch_location_history(self) -> None:
+        """React to the location history being updated."""
+        self._refresh_suggestions()
+
+    def _watch_navigation_history(self) -> None:
+        """React to the navigation history being updated."""
+        self._refresh_suggestions()
+
+    def _watch_bookmarks(self) -> None:
+        """React to the bookmarks being updated."""
+        self._refresh_suggestions()
 
     def _watch_working(self) -> None:
         """React to the working state being updated."""