+++ ChangeLog.md
@@ -8,7 +8,7 @@
   location as shown in the viewer panel, when it is wider than the viewer.
   ([#131](https://github.com/davep/rogallo/pull/131))
 - The command line suggestions now also include all hosts found in the trust
-  store. ([#133](https://github.com/davep/rogallo/pull/133))
+  store. ([#132](https://github.com/davep/rogallo/pull/132))
 
 ## v0.6.0